+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Adds sidekiq-scheduler as a dependency
+- Sets up the schedule of heartbeats within the adapter.
+
 ## [0.7.0] - 2019-07-29
 
 ### Changed